backend/requirements.txt
google-genai
fastapi
uvicorn
python-multipart
pillow
numpy
requests
firebase-admin
